refactor: make protocol conversions explicit

+++ firmware/gps.h
@@ -54,7 +54,7 @@ uint8_t gps_get_fix(struct gps_fix *fix);
 uint8_t gps_wake_up(void);
 void gps_startup_delay(void);
 void delay_ms(unsigned long ms);
-void uart_power_save(int on);
+void uart_power_save(uint8_t on);
